Extract pair_context module from controller.py (#394 Phase 1C)
Move PairContext data class, validate_config(), and configure_lftp() from
controller.py to controller/pair_context.py. Also moves ControllerError.

- PairContext: renamed from _PairContext (no longer private to controller)
- validate_config: takes Context, pure validation function
- configure_lftp: takes Lftp + Config, pure configuration function
- ControllerError: moved alongside PairContext (used by validate_config)
- Controller._validate_config and _configure_lftp become one-line delegates
- Removed unused ScannerResult import from controller.py

controller.py: 1433 → 1180 lines (-253, cumulative with 1A+1B)
New file: controller/pair_context.py (159 lines)
All 163 relevant tests pass unchanged.
