+++ src/python/controller/controller.py
@@ -27,66 +27,10 @@ from .exclude_patterns import filter_excluded_files, parse_exclude_patterns
 from .extract import ExtractProcess, ExtractRequest, ExtractStatus, ExtractStatusResult
 from .model_builder import ModelBuilder
 from .move import MoveProcess
+from .pair_context import ControllerError, PairContext, configure_lftp, validate_config
 from .persist_keys import KEY_SEP, persist_key, strip_persist_key
-from .scan import ActiveScanner, LocalScanner, RemoteScanner, ScannerProcess, ScannerResult
+from .scan import ActiveScanner, LocalScanner, RemoteScanner, ScannerProcess
 from .validate import ValidateProcess, ValidateRequest, ValidateStatusResult
-
-
-class ControllerError(AppError):
-    """
-    Exception indicating a controller error
-    """
-
-    pass
-
-
-class _PairContext:
-    """
-    Holds all per-pair state: LFTP instance, scanners, scanner processes,
-    model builder, and download tracking.
-    """
-
-    def __init__(
-        self,
-        pair_id: str | None,
-        name: str,
-        remote_path: str,
-        local_path: str,
-        effective_local_path: str,
-        lftp: Lftp,
-        active_scanner: ActiveScanner,
-        local_scanner: LocalScanner,
-        remote_scanner: RemoteScanner,
-        active_scan_process: ScannerProcess,
-        local_scan_process: ScannerProcess,
-        remote_scan_process: ScannerProcess,
-        model_builder: ModelBuilder,
-    ):
-        self.pair_id = pair_id
-        self.name = name
-        self.remote_path = remote_path
-        self.local_path = local_path
-        self.effective_local_path = effective_local_path
-        self.lftp = lftp
-        self.active_scanner = active_scanner
-        self.local_scanner = local_scanner
-        self.remote_scanner = remote_scanner
-        self.active_scan_process = active_scan_process
-        self.local_scan_process = local_scan_process
-        self.remote_scan_process = remote_scan_process
-        self.model_builder = model_builder
-
-        # Per-pair tracking state
-        self.active_downloading_file_names: list[str] = []
-        self.active_extracting_file_names: list[str] = []
-        self.prev_downloading_file_names: set[str] = set()
-        self.pending_completion: set[str] = set()
-        self.remote_scan_received: bool = False
-        self.local_scan_received: bool = False
-
-        # Temporary storage for latest scan results (set during _update_pair_model_state)
-        self._latest_remote_scan: ScannerResult | None = None
-        self._latest_local_scan: ScannerResult | None = None
 
 
 class Controller:
@@ -171,7 +115,7 @@ class Controller:
         self.__mp_logger = MultiprocessingLogger(self.logger)
 
         # Build pair contexts, then seed each builder with filtered persist state
-        self.__pair_contexts: list[_PairContext] = self._build_pair_contexts()
+        self.__pair_contexts: list[PairContext] = self._build_pair_contexts()
         self._sync_persist_to_all_builders()
 
         # Setup extract process (global -- extraction is local-only)
@@ -198,59 +142,11 @@ class Controller:
         self.__started = False
 
     def _validate_config(self) -> None:
-        """Validate that all required config fields are set (non-None) at startup.
+        validate_config(self.__context)
 
-        Collects all missing fields and raises a single ControllerError listing them.
+    def _build_pair_contexts(self) -> list[PairContext]:
         """
-        missing: list[str] = []
-        config = self.__context.config
-
-        # Lftp required fields
-        lftp_fields = [
-            "remote_address",
-            "remote_username",
-            "remote_port",
-            "remote_path_to_scan_script",
-            "use_ssh_key",
-            "use_temp_file",
-            "num_max_parallel_downloads",
-            "num_max_parallel_files_per_download",
-            "num_max_connections_per_root_file",
-            "num_max_connections_per_dir_file",
-            "num_max_total_connections",
-        ]
-        for field in lftp_fields:
-            if getattr(config.lftp, field) is None:
-                missing.append(f"Lftp.{field}")
-
-        # Controller required fields
-        controller_fields = [
-            "interval_ms_remote_scan",
-            "interval_ms_local_scan",
-            "interval_ms_downloading_scan",
-        ]
-        for field in controller_fields:
-            if getattr(config.controller, field) is None:
-                missing.append(f"Controller.{field}")
-
-        # General required fields
-        if config.general.verbose is None:
-            missing.append("General.verbose")
-
-        # AutoQueue required fields
-        if config.autoqueue.auto_delete_remote is None:
-            missing.append("AutoQueue.auto_delete_remote")
-
-        # Args required fields
-        if self.__context.args.local_path_to_scanfs is None:
-            missing.append("Args.local_path_to_scanfs")
-
-        if missing:
-            raise ControllerError("Required config fields are not set: {}".format(", ".join(missing)))
-
-    def _build_pair_contexts(self) -> list[_PairContext]:
-        """
-        Build a _PairContext for each configured path pair.
+        Build a PairContext for each configured path pair.
         If no path pairs are configured, create a single default pair from config.lftp
         for backward compatibility.
         """
@@ -287,7 +183,7 @@ class Controller:
             ]
 
         self.__context.status.controller.no_enabled_pairs = False
-        contexts: list[_PairContext] = []
+        contexts: list[PairContext] = []
         for pair in enabled_pairs:
             contexts.append(
                 self._create_pair_context(
@@ -296,9 +192,9 @@ class Controller:
             )
         return contexts
 
-    def _create_pair_context(self, pair_id: str | None, name: str, remote_path: str, local_path: str) -> _PairContext:
+    def _create_pair_context(self, pair_id: str | None, name: str, remote_path: str, local_path: str) -> PairContext:
         """
-        Create a fully wired _PairContext with its own LFTP, scanners, and model builder.
+        Create a fully wired PairContext with its own LFTP, scanners, and model builder.
         """
         pair_label = name or pair_id or "default"
         pair_logger = self.logger.getChild(f"Pair[{pair_label}]")
@@ -375,7 +271,7 @@ class Controller:
         model_builder.set_corrupt_files(set())
         model_builder.set_auto_delete_remote(bool(self.__context.config.autoqueue.auto_delete_remote))  # type: ignore[arg-type]
 
-        return _PairContext(
+        return PairContext(
             pair_id=pair_id,
             name=name,
             remote_path=remote_path,
@@ -392,38 +288,7 @@ class Controller:
         )
 
     def _configure_lftp(self, lftp: Lftp):
-        """Apply shared LFTP configuration settings."""
-        cfg = self.__context.config.lftp
-        lftp.num_parallel_jobs = cfg.num_max_parallel_downloads  # type: ignore[assignment]
-        lftp.num_parallel_files = cfg.num_max_parallel_files_per_download  # type: ignore[assignment]
-        lftp.num_connections_per_root_file = cfg.num_max_connections_per_root_file  # type: ignore[assignment]
-        lftp.num_connections_per_dir_file = cfg.num_max_connections_per_dir_file  # type: ignore[assignment]
-        lftp.num_max_total_connections = cfg.num_max_total_connections  # type: ignore[assignment]
-        lftp.use_temp_file = cfg.use_temp_file  # type: ignore[assignment]
-        lftp.temp_file_name = "*" + Constants.LFTP_TEMP_FILE_SUFFIX
-        if cfg.net_limit_rate:
-            lftp.rate_limit = cfg.net_limit_rate
-        if cfg.net_socket_buffer:
-            lftp.net_socket_buffer = cfg.net_socket_buffer
-        if cfg.pget_min_chunk_size:
-            lftp.min_chunk_size = cfg.pget_min_chunk_size
-        if cfg.mirror_parallel_directories is not None:
-            lftp.mirror_parallel_directories = cfg.mirror_parallel_directories
-        if cfg.net_timeout is not None:
-            lftp.net_timeout = cfg.net_timeout
-        if cfg.net_max_retries is not None:
-            lftp.net_max_retries = cfg.net_max_retries
-        if cfg.net_reconnect_interval_base is not None:
-            lftp.net_reconnect_interval_base = cfg.net_reconnect_interval_base
-        if cfg.net_reconnect_interval_multiplier is not None:
-            lftp.net_reconnect_interval_multiplier = cfg.net_reconnect_interval_multiplier
-        # Configure inline transfer verification
-        validate_cfg = self.__context.config.validate
-        if validate_cfg.xfer_verify:
-            lftp.xfer_verify = True
-            lftp.xfer_verify_command = f"{validate_cfg.algorithm}sum"
-        else:
-            lftp.xfer_verify = False
+        configure_lftp(lftp, self.__context.config)
         lftp.set_verbose_logging(self.__context.config.general.verbose)  # type: ignore[arg-type]
 
     def start(self):
@@ -548,7 +413,7 @@ class Controller:
             model_files.append(copy.deepcopy(file))
         return model_files
 
-    def _get_pair_context_for_command(self, command: Command) -> _PairContext | None:
+    def _get_pair_context_for_command(self, command: Command) -> PairContext | None:
         """Find the pair context for a command based on pair_id."""
         if command.pair_id:
             for pc in self.__pair_contexts:
@@ -557,14 +422,14 @@ class Controller:
             return None
         return self.__pair_contexts[0] if self.__pair_contexts else None
 
-    def _get_pair_context_for_file(self, file: ModelFile) -> _PairContext | None:
+    def _get_pair_context_for_file(self, file: ModelFile) -> PairContext | None:
         """Find the pair context that owns a ModelFile based on its pair_id."""
         for pc in self.__pair_contexts:
             if pc.pair_id == file.pair_id:
                 return pc
         return None
 
-    def _find_pair_by_id(self, pair_id: str | None) -> _PairContext | None:
+    def _find_pair_by_id(self, pair_id: str | None) -> PairContext | None:
         """Find the pair context by pair_id.
         Returns default (first) pair when pair_id is None.
         Returns None when pair_id is provided but not found.
@@ -576,7 +441,7 @@ class Controller:
             return None
         return self.__pair_contexts[0] if self.__pair_contexts else None
 
-    def _build_extract_request(self, file: ModelFile, pc: _PairContext) -> ExtractRequest:
+    def _build_extract_request(self, file: ModelFile, pc: PairContext) -> ExtractRequest:
         """Build an ExtractRequest with the correct pair-specific paths."""
         # Determine output directory
         if self.__context.config.controller.use_local_path_as_extract_path:
@@ -887,7 +752,7 @@ class Controller:
 
     def _update_pair_model_state(  # noqa: C901 — will be decomposed in #394
         self,
-        pc: _PairContext,
+        pc: PairContext,
         latest_extract_statuses: ExtractStatusResult | None,
         latest_validate_statuses: ValidateStatusResult | None,
     ) -> None:
@@ -1122,7 +987,7 @@ class Controller:
                 process = DeleteLocalProcess(local_path=delete_path, file_name=file.name)
                 process.set_mp_log_queue(self.__mp_logger.queue, self.__mp_logger.log_level)
 
-                def post_callback(delete_path: str = delete_path, _pc: _PairContext = pc) -> None:
+                def post_callback(delete_path: str = delete_path, _pc: PairContext = pc) -> None:
                     _pc.local_scan_process.force_scan()
                     if delete_path != _pc.local_path:
                         _pc.active_scan_process.force_scan()
@@ -1251,7 +1116,7 @@ class Controller:
             return
         self.__spawn_move_process(file_name, pc)
 
-    def __spawn_move_process(self, file_name: str, pc: _PairContext):
+    def __spawn_move_process(self, file_name: str, pc: PairContext):
         """
         Spawn a MoveProcess to move a file from staging to the final local_path
         """
